+++ src/ts/ui/setContentTheme.ts
@@ -1,6 +1,9 @@
 import {addStyle} from "../util/addStyle";
 
 export const setContentTheme = (contentTheme: string, path: string) => {
+    if (!contentTheme || !path) {
+        return;
+    }
     const vditorContentTheme = document.getElementById("vditorContentTheme") as HTMLLinkElement;
     const cssPath = `${path}/${contentTheme}.css`;
     if (!vditorContentTheme) {